requirements.txt
Flask==0.10.1
gevent==1.1rc3
-e git+https://github.com/Nekmo/gevent-websocket.git@python3-support#egg=gevent-websocket-0.9.5
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
six==1.10.0
Werkzeug==0.11.3
wheel==0.24.0
bcrypt==2.0.0
hachoir3-superdesk==3.0a1.post2
pyinotify==0.9.6
blinker==1.4